terminado uptade

+++ app/Controllers/LibroController.php
@@ -76,23 +76,24 @@ class LibroController extends Controller{
         $validacion = $this->validate([
             'imagen' => [
                 'uploaded[imagen]',
-                'file_size[imagen,10240]',
+                'max_size[imagen,1024]',
                 'mime_in[imagen,image/jpg,image/jpeg,image/png]'
             ]
         ]);
         if($validacion){
-            $datosLibro= $model->where('id', $id)->first();
-            $rutaImagen = '../public/uploads/'.$datosLibro['imagen'];
-            unlink($rutaImagen);
-            
+            if($imagen = $this->request->getFile('imagen')){
 
-            $imagen = $this->request->getFile('imagen');
-            $nombreRandom = $imagen->getRandomName();
-            $imagen->move('../public/uploads', $nombreRandom);
-            $data=[
-                'imagen' => $nombreRandom
-            ];
-            $model->update($id, $data);
+                $datosLibro= $model->where('id', $id)->first();
+                $rutaImagen = '../public/uploads/'.$datosLibro['imagen'];
+                unlink($rutaImagen);
+
+                $nombreRandom = $imagen->getRandomName();
+                $imagen->move('../public/uploads', $nombreRandom);
+                $data=[
+                    'imagen' => $nombreRandom
+                ];
+                $model->update($id, $data);
+            }
         }
         return redirect()->to('listar');
